+++ src/main/java/com/springhttpserver/Controllers/UserController.java
@@ -1,4 +1,0 @@
-package com.springhttpserver.Controllers;
-
-public class UserController {
-}