+++ src/main/java/com/springhttpserver/Database/ConnectionType.java
@@ -1,0 +1,6 @@
+package com.springhttpserver.Database;
+
+public enum ConnectionType {
+    hibernate,
+    jdbc
+};