+++ src/main/java/com/springhttpserver/Database/DriverType.java
@@ -1,0 +1,5 @@
+package com.springhttpserver.Database;
+
+public enum DriverType {
+    JDBC
+}