+++ src/main/java/com/springhttpserver/Database/JDBCConnection.java
@@ -1,0 +1,4 @@
+package com.springhttpserver.Database;
+
+public class JDBCConnection {
+}