+++ src/main/java/com/springhttpserver/Models/AppSettings/DatabaseSettings.java
@@ -5,11 +5,14 @@ import org.springframework.stereotype.Component;
 @Component
 public class DatabaseSettings {
 
+    public int MaxDBConnections;
     public String Endpoint;
     public int Port;
     public String DbName;
     public String Username;
     public String Password;
+
+    public String DriverType;
 
     public DatabaseSettings()
     {
@@ -18,5 +21,8 @@ public class DatabaseSettings {
         this.DbName = System.getProperty("aws_database.rds_name");
         this.Username = System.getProperty("aws_database.rds_username");
         this.Password = System.getProperty("aws_database.rds_password");
+        this.MaxDBConnections = Integer.parseInt(System.getProperty("aws_database.max_connections"));
+
+        this.DriverType = System.getProperty("mysql.driver");
     }
 }