+++ src/main/java/com/springhttpserver/Services/IHelloService.java
@@ -1,8 +1,0 @@
-package com.springhttpserver.Services;
-
-import com.springhttpserver.Models.HelloData;
-
-public interface IHelloService {
-    String GetGreetingMessage();
-    HelloData GetUserData(String userId);
-}