+++ src/main/java/com/springhttpserver/Services/IUserProfileService.java
@@ -1,0 +1,9 @@
+package com.springhttpserver.Services;
+
+import com.springhttpserver.Models.UserProfile;
+
+public interface IUserProfileService {
+    UserProfile GetUserProfile(String userId);
+    UserProfile CreateNewUserProfile(UserProfile userProfile);
+    boolean DeleteUserProfile(String userId);
+}